Create root page with a management command
Migrations did not work properly, because

    ./manage.py migrate --noinput

did not create a user in auth, and it was not therefore not possible to
create the root page, because this needs a user with id=1 to exist. The
creation of the root page has therefore moved from a migration to a
separate management command.